test(cli): extract normal output boundary helper
Refs: openspec/changes/civ7-support-direct-control-modularization

Move the compact play projection debug/internal marker omission assertion into packages/cli/test/commands/game/play/normal-output-boundary.ts and reuse it from priorities, ready-city, and unit-move-preview coverage.

Record this as local CLI test-helper ownership for the Debug/Internal Service Output proof lane only. This does not accept Task 2.9.4, assign production source owners, claim runtime/live-game proof, implement semantic envelopes, create telemetry or AI-ingestion contracts, or unblock Effect/oRPC procedure-core work.

Validation:
- bun run --cwd packages/cli test -- test/commands/game/play/priorities.test.ts test/commands/game/play/ready-city.test.ts test/commands/game/play/unit-move-preview.test.ts
- bun run check:cli
- bun run test:cli:play
- bun run test:cli -- --force
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- bun run openspec:validate
- git diff --check
