+++ openspec/changes/civ7-support-direct-control-modularization/tasks.md
@@ -118,8 +118,9 @@
         `workstream/debug-service-projection-contract.md`, but it does not
         assign source/proof owners or accept the row. Focused compact
         `game play priorities`, compact `game play ready-city`, and compact
-        `game play unit-move-preview` coverage now prove three normal play
-        projections omit raw
+        `game play unit-move-preview` coverage, now sharing the
+        `game/play/normal-output-boundary.ts` test helper, prove three normal
+        play projections omit raw
         transport/session/probe/correlation command internals, but broader
         debug-command boundary, AI-ingestion, telemetry, and
         procedure-diagnostic separation proof remains pending before row
